requirements.txt
exceptiongroup==1.1.3
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
pytest==7.4.1
python-dotenv==1.0.0
tomli==2.0.1
